requirements.txt
camelot-py==1.0.0
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.3
distro==1.9.0
dotenv==0.9.9
et_xmlfile==2.0.0
jpype1==1.5.2
numpy==2.2.6
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
pycparser==2.22
pycryptodome==3.23.0
pypdf==5.6.0
PyPDF2==3.0.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
tabula-py==2.10.0
tabulate==0.9.0
tzdata==2025.2
